porting to NETFX_CORE...

+++ Packet.cs
@@ -3,16 +3,16 @@
     public class Packet
     {
         private int peerId;
-        private uint seq;
+        private byte seq;
         private byte[] payload;
 
 
         public int PeerId { get { return peerId; } }
-        public uint Seq { get { return seq; } }
+        public byte Seq { get { return seq; } }
         public byte[] Data { get { return payload; } }
 
 
-        public Packet(int peerId, uint seq, byte[] data) // TODO only of ip or peerId need be supplied then the other is calculated
+        public Packet(int peerId, byte seq, byte[] data) // TODO only of ip or peerId need be supplied then the other is calculated
         {
             this.peerId = peerId;
             this.seq = seq;
