+++ Settings.cs
@@ -14,6 +14,6 @@ namespace FalconUDP
         internal const int ACKTickTime                  = 1000;                             // milliseconds (note timer could tick just after packet sent so this also defines the error margin)
         internal const int ACKTimeoutTicks              = ACKTimeout / ACKTickTime;         // timeout in ticks 
         internal const int ACKRetryAttempts             = 3;                                // number of times to retry until assuming peer is dead (used by AwaitingAcceptDetail too)
-        internal const byte OutOfOrderTolerance         = (byte)Const.HALF_MAX_SEQ_NUMS / 2;      // packets recveived out-of-order from last received greater than this are dropped
+        internal const int OutOfOrderTolerance          = Const.HALF_MAX_SEQ_NUMS / 2;      // packets recveived out-of-order from last received greater than this are dropped
     }
 }
